+++ packages/lytenyte-shared/src/renderer-react/__tests__/use-header-cell-style.test.tsx
@@ -15,8 +15,9 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
+        "gridColumnStart": "1",
+        "overflow": "hidden",
         "position": "relative",
-        "transform": "translate3d(0px, 0px, 0px)",
       }
     `);
   });
@@ -33,9 +34,10 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
-        "left": 0,
+        "gridColumnStart": "1",
+        "insetInlineStart": "0px",
+        "overflow": "hidden",
         "position": "sticky",
-        "transform": "translate3d(0px, 0px, 0px)",
         "zIndex": 11,
       }
     `);
@@ -53,9 +55,10 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
-        "left": 0,
+        "gridColumnStart": "2",
+        "insetInlineEnd": "200px",
+        "overflow": "hidden",
         "position": "sticky",
-        "transform": "translate3d(700px, 0px, 0px)",
         "zIndex": 11,
       }
     `);
@@ -73,8 +76,9 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
+        "gridColumnStart": "1",
+        "overflow": "hidden",
         "position": "relative",
-        "transform": "translate3d(0px, 0px, 0px)",
       }
     `);
   });
@@ -91,9 +95,10 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
+        "gridColumnStart": "1",
+        "insetInlineStart": "0px",
+        "overflow": "hidden",
         "position": "sticky",
-        "right": 0,
-        "transform": "translate3d(0px, 0px, 0px)",
         "zIndex": 11,
       }
     `);
@@ -111,9 +116,10 @@ describe("useHeaderCellStyle", () => {
 
     expect(result.result.current).toMatchInlineSnapshot(`
       {
+        "gridColumnStart": "2",
+        "insetInlineEnd": "200px",
+        "overflow": "hidden",
         "position": "sticky",
-        "right": 0,
-        "transform": "translate3d(-700px, 0px, 0px)",
         "zIndex": 11,
       }
     `);