+++ packages/commercial-lytenyte-license/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @1771technologies/lng-license
 
+## 1.0.12-dev.2
+
+### Patch Changes
+
+- Fix bad compile
+
+## 1.0.12-dev.1
+
+### Patch Changes
+
+- Column fixes
+
+## 1.0.12-dev.0
+
+### Patch Changes
+
+- Dev version
+
 ## 1.0.11
 
 ### Patch Changes