+++ packages/dom-utils/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @1771technologies/lytenyte-dom-utils
 
+## 1.0.12-dev.2
+
+### Patch Changes
+
+- Fix bad compile
+- Updated dependencies
+  - @1771technologies/lytenyte-js-utils@1.0.12-dev.2
+
+## 1.0.12-dev.1
+
+### Patch Changes
+
+- Column fixes
+- Updated dependencies
+  - @1771technologies/lytenyte-js-utils@1.0.12-dev.1
+
+## 1.0.12-dev.0
+
+### Patch Changes
+
+- Dev version
+- Updated dependencies
+  - @1771technologies/lytenyte-js-utils@1.0.12-dev.0
+
 ## 1.0.11
 
 ### Patch Changes