+++ packages/play-frame/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @1771technologies/play-frame
 
+## 0.0.11-dev.2
+
+### Patch Changes
+
+- Fix bad compile
+
+## 0.0.11-dev.1
+
+### Patch Changes
+
+- Column fixes
+
+## 0.0.11-dev.0
+
+### Patch Changes
+
+- Dev version
+
 ## 0.0.10
 
 ### Patch Changes